Update docker_singularity_intro.html
I changed the maroon colors to navy. I also updated the <tt> tags to <code> tags. I read that <tt> is deprecated and not supported universally, but I don't know if it was necessary for me to change or not. There was an issue ending a <tt> code segment with <\tt> resulting in everything after a certain chunk being in code.