Added email info and fixed bugs
The hostname can't be bound to a variable and passed to MPI since the binding will happen before mpi runs the remote command. Changed the default time to 2 hours instead of 48 since a lot of users assume 48 hours is the default time they should use based on this tutorial. Added email notifications and removed the shell information. Email notifications are used far more often than changing the shell.