Replaced keepbothreads with True
I wanted to reach out to CARC to let you know that there is an error in one of the Quickbytes for GATK on your website.
For the trimming reads portion of the pipeline, there is a line for trimmomatic that is:
ILLUMINACLIP:${adapters}:2:30:10:2:keepBothReads
The keep both reads flag needs to be “True”:
ILLUMINACLIP:${adapters}:2:30:10:2:True

This is really important because without the flag as “true”, it throws out a substantial amount of read data. A collaborator of mine found this error for me, and we had to redo 100 genomes. It’s the difference between 59 low-quality resequenced genomes having an average read depth of <1X (when it was “keepBothReads” to ~8X (when it was changed to “True”).
Thought this might of help for future researchers! Regardless of this issue, the Quickbytes are a huge resource for me. I’m very happy that CARC has produced these public tutorials.
Jenna McCullough